requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
emails==0.6
pydantic==2.12.4
email-validator==2.2.0
